erp test aşaması

+++ Erp.Web/Controllers/CariController.cs
@@ -1,15 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Erp.Data;
 using Erp.Data.Entities;
-using Microsoft.AspNetCore.Authorization;
-
-
 
 namespace Erp.Web.Controllers
 {
@@ -30,19 +23,13 @@ namespace Erp.Web.Controllers
         }
 
         // GET: Cari/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(string? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var cari = await _context.Cariler
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (cari == null)
-            {
-                return NotFound();
-            }
+                .FirstOrDefaultAsync(m => m.CariKodu == id);
+            if (cari == null) return NotFound();
 
             return View(cari);
         }
@@ -54,12 +41,29 @@ namespace Erp.Web.Controllers
         }
 
         // POST: Cari/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,CariKodu,CariUnvani,CariKisaAd,CariTipi,Durumu,VergiDairesi,VergiNo,TCKN,MersisNo,TicaretSicilNo,YetkiliKisi,Telefon,EPosta,WebSitesi,Ulke,Il,Ilce,PostaKodu,Adres,SatisTemsilcisi,Bolge,Sektor,DovizTuru,RiskLimiti,OlusturmaTarihi,GuncellemeTarihi")] Cari cari)
+        public async Task<IActionResult> Create([Bind("CariKodu,CariUnvani,CariKisaAd,CariTipi,Durumu,VergiDairesi,VergiNo,TCKN,MersisNo,TicaretSicilNo,YetkiliKisi,Telefon,EPosta,WebSitesi,Ulke,Il,Ilce,PostaKodu,Adres,SatisTemsilcisi,Bolge,Sektor,DovizTuru,RiskLimiti,OlusturmaTarihi,GuncellemeTarihi")] Cari cari)
         {
+            if (string.IsNullOrEmpty(cari.CariKodu))
+            {
+                var sonCari = _context.Cariler
+                    .OrderByDescending(c => c.CariKodu)
+                    .FirstOrDefault();
+
+                int sonNumara = 0;
+                if (sonCari != null && sonCari.CariKodu.StartsWith("C-"))
+                {
+                    var numaraStr = sonCari.CariKodu.Replace("C-", "");
+                    int.TryParse(numaraStr, out sonNumara);
+                }
+
+                int yeniNumara = sonNumara + 1;
+                cari.CariKodu = $"C-{yeniNumara:D4}";
+            }
+
+            ModelState.Remove("CariKodu");
+
             if (ModelState.IsValid)
             {
                 _context.Add(cari);
@@ -70,50 +74,34 @@ namespace Erp.Web.Controllers
         }
 
         // GET: Cari/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Edit(string? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var cari = await _context.Cariler.FindAsync(id);
-            if (cari == null)
-            {
-                return NotFound();
-            }
+            if (cari == null) return NotFound();
             return View(cari);
         }
 
         // POST: Cari/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,CariKodu,CariUnvani,CariKisaAd,CariTipi,Durumu,VergiDairesi,VergiNo,TCKN,MersisNo,TicaretSicilNo,YetkiliKisi,Telefon,EPosta,WebSitesi,Ulke,Il,Ilce,PostaKodu,Adres,SatisTemsilcisi,Bolge,Sektor,DovizTuru,RiskLimiti,OlusturmaTarihi,GuncellemeTarihi")] Cari cari)
+        public async Task<IActionResult> Edit(string id, [Bind("CariKodu,CariUnvani,CariKisaAd,CariTipi,Durumu,VergiDairesi,VergiNo,TCKN,MersisNo,TicaretSicilNo,YetkiliKisi,Telefon,EPosta,WebSitesi,Ulke,Il,Ilce,PostaKodu,Adres,SatisTemsilcisi,Bolge,Sektor,DovizTuru,RiskLimiti,OlusturmaTarihi,GuncellemeTarihi")] Cari cari)
         {
-            if (id != cari.Id)
-            {
-                return NotFound();
-            }
+            if (id != cari.CariKodu) return NotFound();
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    cari.GuncellemeTarihi = DateTime.Now;
                     _context.Update(cari);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!CariExists(cari.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!CariExists(cari.CariKodu)) return NotFound();
+                    throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -121,19 +109,13 @@ namespace Erp.Web.Controllers
         }
 
         // GET: Cari/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(string? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var cari = await _context.Cariler
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (cari == null)
-            {
-                return NotFound();
-            }
+                .FirstOrDefaultAsync(m => m.CariKodu == id);
+            if (cari == null) return NotFound();
 
             return View(cari);
         }
@@ -141,26 +123,18 @@ namespace Erp.Web.Controllers
         // POST: Cari/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(string id)
         {
             var cari = await _context.Cariler.FindAsync(id);
-            if (cari != null)
-            {
-                _context.Cariler.Remove(cari);
-            }
+            if (cari != null) _context.Cariler.Remove(cari);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool CariExists(int id)
+        private bool CariExists(string id)
         {
-            return _context.Cariler.Any(e => e.Id == id);
-        }
-        public IActionResult Test()
-        {
-            var count = _context.Cariler.Count();
-            return Content($"Cari tablosunda {count} kayıt var!");
+            return _context.Cariler.Any(e => e.CariKodu == id);
         }
     }
 }
